pricing-service/requirements.txt
# Python Pricing ML Service Dependencies
# Simplified without heavy ML libs to avoid build issues on Windows

# Web Framework
fastapi
uvicorn[standard]
pydantic

# Data manipulation
numpy
pandas
pyarrow  # For efficient parquet storage

# Machine Learning
lightgbm
scikit-learn
scipy

# HTTP Client
httpx

# Environment Variables
python-dotenv

# CORS Support
python-multipart

# Testing
pytest
pytest-asyncio
pytest-cov

# Observability
sentry-sdk[fastapi]
prometheus-client

# gRPC
grpcio
grpcio-tools
protobuf
